requirements.txt
openai>=1.18.0
langchain>=0.1.13
python-dotenv>=1.0.1
typer[all]>=0.9.0
pydantic>=2.6.4
rich>=13.7.1
pyyaml>=6.0.1